modified migrate files
