+++ alembic/versions/7d7502ffee48_create_waiting_pool_table.py
@@ -1,8 +1,8 @@
-"""parking space pool
+"""waiting pool
 
-Revision ID: bced7c31fc49
-Revises: fa89180e982f
-Create Date: 2016-11-09 00:28:12.512007
+Revision ID: 7d7502ffee48
+Revises: bced7c31fc49
+Create Date: 2016-11-09 00:33:32.326960
 
 """
 from alembic import op
@@ -10,25 +10,26 @@ import sqlalchemy as sa
 
 
 # revision identifiers, used by Alembic.
-revision = 'bced7c31fc49'
-down_revision = 'fa89180e982f'
+revision = '7d7502ffee48'
+down_revision = 'bced7c31fc49'
 branch_labels = None
 depends_on = None
 
 
 def upgrade():
     op.create_table(
-        'parking_space_pool',
-        sa.Column('id', sa.Integer, primary_key=True, autoincrement=False),
+        'waiting_pool',
         sa.Column('user_id', sa.String(64), sa.ForeignKey('users.user_id'), nullable=False),
         sa.Column('longitude', sa.Float, nullable=False),
         sa.Column('latitude', sa.Float, nullable=False),
         sa.Column('location', sa.String(255), nullable=True),
+        sa.Column('level', sa.Integer, nullable=True),
+        sa.Column('is_active', sa.Boolean, nullable=False, default=False),
         sa.Column('created_at', sa.DateTime, nullable=False, default=sa.func),
-        sa.Column('updated_at', sa.DateTime, nullable=False, default=sa.func),
     )
 
 
 def downgrade():
-    op.drop_table('parking_space_pool')
+    op.drop_table('waiting_pool')
+
