+++ alembic/versions/e6def7ff3050_create_matched_parking_space_list_table.py
@@ -1,4 +1,4 @@
-"""create prereserve list table
+"""create matched parking space list table
 
 Revision ID: e6def7ff3050
 Revises: 7d7502ffee48
@@ -18,15 +18,14 @@ depends_on = None
 
 def upgrade():
     op.create_table(
-        'prereserved_list',
+        'matched_parking_space_list',
         sa.Column('user_id', sa.String(64), sa.ForeignKey('users.user_id'), nullable=False),
-        # TODO add foreign key constraint (ParkingSpace.id)
-        sa.Column('parking_id', sa.Integer, nullable=False),
+        sa.Column('plate', sa.String(7), sa.ForeignKey('vehicles.plate'), nullable=False),
+        sa.Column('status', sa.Enum('awaiting', 'rejected', 'accepted', 'expired', name='status_types'),
+                  nullable=False, default='awaiting'),
         sa.Column('created_at', sa.DateTime, nullable=False, default=sa.func),
-        sa.Column('updated_at', sa.DateTime, nullable=False, default=sa.func),
     )
 
-
 def downgrade():
-    op.drop_table('prereserved_list')
+    op.drop_table('matched_parking_space_list')
